requirements.txt
torch>=1.6
ptflops~=0.6.8
tqdm~=4.62
numpy
psutil>=5.9
gputil>=1.4
py-cpuinfo>=7.0
pyyaml >= 6.0
carbontracker >= 1.1